:recycle: [Refactor] comment addDialog, editDialog feature의ui로 이동 #main
